hesai_decoders. update timestamp calculation

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/pandar_qt_128_decoder.hpp
@@ -46,7 +46,7 @@ private:
   /// @param unix_second Packet time
   /// @return Point cloud
   drivers::NebulaPoint build_point(
-    size_t block_id, size_t unit_id, bool dual_return, const uint32_t & unix_second);
+    size_t block_id, size_t unit_id, bool dual_return, const double & unix_second);
   /// @brief Convert to point cloud
   /// @param block_id target block
   /// @return Point cloud
